[refac] update package structure for security components
