+++ edukit-api/src/main/java/com/edukit/auth/security/authentication/MemberAuthentication.java
@@ -1,4 +1,4 @@
-package com.edukit.common.security.authentication;
+package com.edukit.auth.security.authentication;
 
 import java.util.Collection;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
