+++ edukit-api/src/main/java/com/edukit/auth/security/authentication/MemberDetailReader.java
@@ -1,4 +1,4 @@
-package com.edukit.common.security.authentication;
+package com.edukit.auth.security.authentication;
 
 import com.edukit.core.member.db.entity.Member;
 import com.edukit.core.member.db.enums.MemberRole;
