+++ edukit-api/src/main/java/com/edukit/auth/security/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.edukit.common.security.config;
+package com.edukit.auth.security.config;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Value;
