+++ edukit-api/src/main/java/com/edukit/auth/security/handler/RefreshTokenCookieHandler.java
@@ -1,4 +1,4 @@
-package com.edukit.common.security.handler;
+package com.edukit.auth.security.handler;
 
 import com.edukit.core.auth.service.jwt.setting.JwtProperties;
 import java.time.Duration;
